refactor: domain 내의 패키지 분리
결제, 주문, 상품을 의미하는 payment, order, stock으로 패키지를 분리한다.
