+++ src/main/java/store/domain/payment/Payment.java
@@ -1,4 +1,7 @@
-package store.domain;
+package store.domain.payment;
+
+import store.domain.order.Order;
+import store.domain.stock.Stock;
 
 public class Payment {
     private final Order order;
