+++ src/main/java/store/domain/stock/Product.java
@@ -1,4 +1,4 @@
-package store.domain;
+package store.domain.stock;
 
 public class Product {
