+++ src/test/java/store/domain/order/OrderTest.java
@@ -1,4 +1,4 @@
-package store.domain;
+package store.domain.order;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
